test: move all tests for methods implemented in abstract binary tree class to separate test class
